tools/sender/requirements.txt
google-auth==2.33.0
requests==2.32.3
